chore: create _lists.py file
